Fix: remove `version` tag from `composer.json`.
Our recent tags are not showing on Packagist because our config file had a
hardcoded version number, and this confuses Composer's understanding of what
version a particular tag is (e.g. is it the tag number, or the hardcoded num?).

There is generally no need to manually track version numbers in a text file for
Composer, so we're simply removing this.