Fix roles in multisite
When the network roles syncing is enabled, there's a bug in the `foreach` trying to iterate over the object (which doesn't implement iterable etc). So far as I can see, this would never have worked.

I also increased the `number` param, else it will default to 100. There is no support for `-1` so I'm not sure what the ideal fix here is, but even in our own case we have networks with > 100 sites.